Introduction de ScheduleColorizer dont la responsabilité déterminer la couleur d'un éléments de planning, permettant de mutualiser le code de détermination dans les vues et l'export (précédemment c'était redondé dans l'export et en retard sur les vues)
