+++ src/net/algem/planning/day/DayScheduleCtrl.java
@@ -35,6 +35,7 @@ import java.util.List;
 import java.util.Locale;
 import javax.swing.*;
 import net.algem.Algem;
+import net.algem.config.ColorPrefs;
 import net.algem.config.ConfigKey;
 import net.algem.config.ConfigUtil;
 import net.algem.planning.*;
@@ -210,7 +211,10 @@ public class DayScheduleCtrl
       File destFile = FileUtil.getSaveFile(view, "xls", BundleUtil.getLabel("Excel.file.label"), null);
       if (destFile != null) {
         try {
-          new PlanningExportService(new PlanningService(DataCache.getDataConnection())).exportPlanning(planning, destFile);
+          new PlanningExportService(
+                  new PlanningService(DataCache.getDataConnection()),
+                  new ScheduleColorizer(new ColorPrefs(), (ActionIO) DataCache.getDao(Model.Action))
+          ).exportPlanning(planning, destFile);
           new DesktopOpenHandler().open(destFile.getAbsolutePath());
         } catch (IOException e) {
           GemLogger.log(e.getMessage());
